[FINISHES #129019501] Reajustando método para o teste de controlador, commit final
